requirements.txt
streamlit
streamlit-authenticator
pandas
python-dotenv
